+++ tools/wsbrd_cli/wsbrddbusapi.rs
@@ -134,7 +134,7 @@ pub trait ComSilabsWisunBorderRouter {
     fn set_udp_dst_port(&self, arg0: u16) -> Result<(), dbus::Error>;
     fn socket_udp_sent_to(&self, arg0: Vec<u8>) -> Result<(), dbus::Error>;
     fn set_multicast_addr(&self, arg0: Vec<u8>) -> Result<(), dbus::Error>;
-    fn set_udp_body_uint_repeat_time(&self, arg0: u8) -> Result<(), dbus::Error>;
+    fn set_udp_body_uint_repeat_time(&self, arg0: u16) -> Result<(), dbus::Error>;
     fn set_udp_tail(&self, arg0: Vec<u8>) -> Result<(), dbus::Error>;
     fn set_udp_body_unit(&self, arg0: Vec<u8>) -> Result<(), dbus::Error>;
     fn set_wisun_gtk_time_settings(&self, arg0: u8, arg1: u8, arg2: u8, arg3: u32) -> Result<(), dbus::Error>;
@@ -337,7 +337,7 @@ impl<'a, T: blocking::BlockingSender, C: ::std::ops::Deref<Target=T>> ComSilabsW
         self.method_call("com.silabs.Wisun.BorderRouter", "setMultcastAddr", (arg0, ))
     }
 
-    fn set_udp_body_uint_repeat_time(&self, arg0: u8) -> Result<(), dbus::Error> {
+    fn set_udp_body_uint_repeat_time(&self, arg0: u16) -> Result<(), dbus::Error> {
         self.method_call("com.silabs.Wisun.BorderRouter", "setUdpBodyUintRepeatTtime", (arg0, ))
     }
